RandomTool.getRandomBalance return BigDecimal
Took 1 minute

+++ src/test/java/pl/wallet/transaction/TransactionTool.java
@@ -4,6 +4,7 @@ import pl.test_tool.RandomTool;
 import pl.wallet.WalletDto;
 import pl.wallet.category.CategoryDto;
 
+import java.math.BigDecimal;
 import java.security.Principal;
 
 public class TransactionTool {
@@ -26,8 +27,8 @@ public class TransactionTool {
     return transactionResource.addTransaction(principal, categoryDto.getId(), walletDto.getId(), getRandomTransaction()).getBody();
   }
 
-  private static Integer getRandomBalance () {
-    return RandomTool.getNumberInteger();
+  private static BigDecimal getRandomBalance () {
+    return BigDecimal.valueOf(RandomTool.getNumberInteger());
   }
 
   private static String getRandomName () {
